backend/requirements.txt
fastapi
uvicorn
python-multipart

httpx
aiofiles
requests
daytona
pydantic
pydantic-settings
python-dotenv

openhands-sdk
openhands-tools
